feat(wasm-visor): embed the wasm-visor by default (committed blob, manual updates)
Per the decision to have `go install github.com/skycoin/skywire@…` carry the
browser visor: the std-Go wasm-visor is now embedded by DEFAULT (no build tag).
pkg/wasmhv/wasmbin/wasm-visor.wasm.gz (~8MB gzipped from ~37MB) is COMMITTED so it
travels in the published Go module, and embed.go go:embeds it unconditionally.
Every build — `make build`, releases, and a plain `go install` — now produces a
skywire whose `cli hv gen` needs no --wasm.

Supersedes the build-tag approach (#3266): the gitignored+`-tags embedwasm` trick
kept history clean but was invisible to `go install` (the blob wasn't in the
module, and `-tags=embedwasm` failed with "no matching files"). Default-embed is
the only way `go install` gets it.

To keep the blob from churning git history:
- `make wasm-visor` builds into build/ only and does NOT touch the committed blob,
  so day-to-day builds / `go test ./...` never dirty it.
- `make embed-wasm-visor` deterministically (gzip -9 -n) rewrites the committed
  blob — run intentionally, commit deliberately. History gains a new ~8MB blob
  only on a deliberate embed bump, not per code change.
- scripts/prune-wasm-embed-history.sh strips old blob versions from history
  (keeps current) for periodic manual cleanup — with the loud caveats that it
  rewrites history (force-push + re-clone) and the Go proxy retains already-
  published versions immutably.

Removed embed_on.go/embed_off.go + the build-embedwasm target. gen.go messages
updated. docs/design/wasm-visor-distribution.md rewritten. Verified: default
`go build` embeds it; wasmbin test decompresses a valid 37MB wasm; `cli hv gen`
works with no --wasm.
