+++ .gitignore
@@ -134,5 +134,3 @@ skysocks-client
 /visor.log
 /hist/
 
-# Embedded wasm-visor binary (built by `make wasm-visor`, embedded under -tags embedwasm; never committed)
-pkg/wasmhv/wasmbin/wasm-visor.wasm.gz
