+++ Makefile
@@ -1,7 +1,7 @@
 
 .PHONY : check lint install-linters dep test lint-extra
 .PHONY : update-deps update-dmsg update-skycoin push-deps
-.PHONY : build clean install format  bin build-race deploy build-embedwasm wasm-visor
+.PHONY : build clean install format  bin build-race deploy wasm-visor embed-wasm-visor
 .PHONY : host-apps bin
 .PHONY : docker-image docker-clean docker-network
 .PHONY : docker-apps docker-bin docker-volume
@@ -239,19 +239,17 @@ tinygo-wasm-visor: ## Build the browser WASM visor edge (dmsg+transport+router+a
 	cp ./pkg/wasmhv/browse.js ./build/wasm-visor/
 	@echo "built ./build/wasm-visor (TinyGo) — serve it: 'go run cmd/dmsg-wasm/serve.go -dir build/wasm-visor' then open http://localhost:8085/"
 
-wasm-visor: ## Build the browser WASM visor edge with STANDARD Go js/wasm into build/wasm-visor-go — larger (~38MB) but full crypto/tls + net/http (https clearnet via skysocks)
+wasm-visor: ## Build the browser WASM visor edge with STANDARD Go js/wasm into build/wasm-visor-go — larger (~38MB) but full crypto/tls + net/http (https clearnet via skysocks). Does NOT touch the committed embed blob.
 	mkdir -p ./build/wasm-visor-go
 	GOOS=js GOARCH=wasm go build -ldflags="-s -w" -o ./build/wasm-visor-go/wasm-visor.wasm ./cmd/wasm-visor
 	cp "$$(go env GOROOT)/lib/wasm/wasm_exec.js" ./build/wasm-visor-go/wasm_exec.js
 	cp ./cmd/wasm-visor/index.html ./build/wasm-visor-go/
 	cp ./pkg/wasmhv/browse.js ./build/wasm-visor-go/
-	gzip -9 -c ./build/wasm-visor-go/wasm-visor.wasm > ./pkg/wasmhv/wasmbin/wasm-visor.wasm.gz
-	@echo "built ./build/wasm-visor-go (standard Go js/wasm) + embed gz at pkg/wasmhv/wasmbin/ (gitignored)"
-	@echo "  serve dev: 'go run cmd/dmsg-wasm/serve.go -dir build/wasm-visor-go'; embed in skywire: 'make build-embedwasm'"
+	@echo "built ./build/wasm-visor-go (standard Go js/wasm) — serve dev: 'go run cmd/dmsg-wasm/serve.go -dir build/wasm-visor-go'"
 
-build-embedwasm: wasm-visor ## Build skywire with the std-Go wasm-visor embedded (so `cli hv gen` needs no --wasm). Runs `make wasm-visor` first.
-	${OPTS} go build -tags embedwasm ${BUILD_OPTS} -o $(BUILD_PATH)skywire .
-	@echo "built $(BUILD_PATH)skywire WITH embedded wasm-visor (~8MB larger) — `cli hv gen` works without --wasm"
+embed-wasm-visor: wasm-visor ## Update the COMMITTED embedded wasm-visor blob (pkg/wasmhv/wasmbin/wasm-visor.wasm.gz) — run intentionally, then `git add` + commit it. Deterministic gzip (-n) so re-running on the same wasm yields no diff.
+	gzip -9 -n -c ./build/wasm-visor-go/wasm-visor.wasm > ./pkg/wasmhv/wasmbin/wasm-visor.wasm.gz
+	@echo "updated pkg/wasmhv/wasmbin/wasm-visor.wasm.gz — review with 'git status', commit intentionally (it's a ~8MB blob)."
 
 dmsg-wasm-hv: ## Build the browser hypervisor-over-dmsg bundle (Service Worker proxy) into build/dmsg-wasm-hv
 	mkdir -p ./build/dmsg-wasm-hv
