+++ docs/design/wasm-visor-distribution.md
@@ -1,58 +1,57 @@
-# wasm-visor binary: embed it with the visor, without git-history accumulation
+# wasm-visor binary: embedded by default, blob updated intentionally
 
-## The problem
+## What ships
 
-The standard-Go `wasm-visor.wasm` (the browser visor, since TinyGo can't do TLS —
-see [tinygo issue #3259]) is **~37 MB** and changes on every build. We want it
-**shipped inside the skywire binary** (so `cli hv gen` — and, later, the visor
-serving the HV UI — produce a browser visor with no external file), but git keeps
-every version of a tracked file forever, so committing it or `go:embed`-ing a
-*tracked* copy would add a fresh multi-MB blob to history on every change.
+The standard-Go `wasm-visor.wasm` (the browser visor — TinyGo can't do TLS, see
+[tinygo issue #3259]) is **embedded in the skywire binary by default**, gzipped:
+`pkg/wasmhv/wasmbin/wasm-visor.wasm.gz` (~8 MB, from the ~37 MB wasm) is
+**committed** and `go:embed`-ed with no build tag. So every build — `make build`,
+the release pipeline, and a plain `go install github.com/skycoin/skywire@…` —
+carries a working wasm-visor, and `skywire cli hv gen` produces a browser visor
+with no `--wasm` flag.
 
-## The approach: build-tagged embed of a gitignored binary
+`cli hv gen` sources the wasm as: `--wasm <path>` if given (e.g. a TinyGo build,
+which also needs `--wasm-exec`) → otherwise the embedded std-Go wasm-visor (uses
+Go's embedded `wasm_exec.js`, no `--wasm-exec`).
 
-The binary is embedded, but the embedded file is **never committed**:
+## Keeping the blob from churning
 
-- `make wasm-visor` builds the wasm and writes a gzip of it to
-  `pkg/wasmhv/wasmbin/wasm-visor.wasm.gz` (~8 MB). That path is `.gitignore`d.
-- `pkg/wasmhv/wasmbin` embeds it **only under the `embedwasm` build tag**
-  (`embed_on.go`: `//go:embed wasm-visor.wasm.gz`). Default builds compile
-  `embed_off.go` instead and carry nothing — so a plain `go build ./...` (CI,
-  lint, dev) neither needs the file nor pays the size.
-- `make build-embedwasm` runs `make wasm-visor` then
-  `go build -tags embedwasm` — this is the build that ships the wasm-visor inside
-  skywire (~8 MB larger).
+The committed `.gz` is **only updated on purpose**:
 
-Because the `.wasm.gz` is gitignored and produced at build time, **git history
-never accumulates the binary** — only the tiny source files in `wasmbin/` are
-tracked. Each build simply overwrites the local gz.
+- `make wasm-visor` builds the wasm into `build/wasm-visor-go/` and does **not**
+  touch the committed blob — so day-to-day builds and `go test ./...` never dirty
+  it.
+- `make embed-wasm-visor` rebuilds and **deterministically** (`gzip -9 -n`,
+  stripping name+mtime) writes the committed `pkg/wasmhv/wasmbin/wasm-visor.wasm.gz`.
+  Run it when you intend to bump the embedded visor, then `git add` + commit that
+  blob deliberately.
 
-Do **not** commit `pkg/wasmhv/wasmbin/wasm-visor.wasm.gz` (it's gitignored for
-this reason) or change the embed to a non-tag-gated `go:embed` of a tracked file.
+So git history gains a new ~8 MB blob only on a deliberate embed update, not on
+every code change.
 
-## Consuming it
+## Pruning old blobs from history (manual, occasional)
 
-`skywire cli hv gen` sources the wasm in this order:
+Updating the embedded wasm replaces the blob; the **previous** versions stay in
+git history and add up over time. To reclaim that space, periodically prune the
+historical versions (keeping the current one) with `git filter-repo`:
 
-1. `--wasm <path>` if given (e.g. a TinyGo build, which also needs `--wasm-exec`);
-2. otherwise the **embedded** std-Go wasm-visor (`wasmbin.Embedded()`), which uses
-   Go's embedded `wasm_exec.js` — no `--wasm-exec` needed;
-3. otherwise an error telling you to pass `--wasm` or build `-tags embedwasm`.
+    scripts/prune-wasm-embed-history.sh      # see the script for the exact steps
 
-So a release-built skywire (`make build-embedwasm`) runs `cli hv gen` with no
-flags. A default-built skywire still works via `--wasm`.
+**Important caveats** — this is a history rewrite, so:
 
-## Release / packaging
-
-Build the shipped skywire with `make build-embedwasm` so the wasm-visor travels
-with the binary. The `.wasm.gz` is a build artifact, not a committed file; the
-release pipeline produces it via `make wasm-visor` as part of `build-embedwasm`.
+- It changes commit hashes → it requires a coordinated **force-push**, and every
+  clone/fork must re-clone or hard-reset. Don't do it casually on a shared branch.
+- The **Go module proxy + sumdb retain already-published versions immutably**.
+  `go install …@v1.3.x` for a version that shipped the blob still fetches that
+  blob from the proxy regardless of pruning. Pruning shrinks the *git repo* and
+  fresh clones, not what the proxy serves for past releases. (One more reason to
+  bump the embedded blob infrequently — each tagged release that changes it is
+  permanent in the proxy.)
 
 ## TODO when TinyGo gains software TLS
 
-Once [tinygo issue #3259] is resolved, the small (~2.2 MB) TinyGo build can do
-HTTPS again and becomes the default embed, shrinking the addition ~4×. The
-mechanism is unchanged: build it, gzip it to the gitignored embed path, embed
-under the tag.
+Once [tinygo issue #3259] lands, the ~2.2 MB TinyGo build can do HTTPS and becomes
+the default embed, cutting the committed blob ~4×. Same mechanism: `make
+embed-wasm-visor`, commit intentionally, prune occasionally.
 
 [tinygo issue #3259]: https://github.com/skycoin/skywire/issues/3259
