+++ pkg/wasmhv/wasmbin/embed_test.go
@@ -1,12 +1,13 @@
-//go:build embedwasm
-
 package wasmbin
 
 import "testing"
 
+// TestEmbeddedGet confirms the committed wasm-visor.wasm.gz decompresses to a
+// valid wasm binary (so a default build — incl. `go install` — really carries a
+// working wasm-visor).
 func TestEmbeddedGet(t *testing.T) {
 	if !Embedded() {
-		t.Fatal("Embedded() false under -tags embedwasm")
+		t.Fatal("wasm-visor binary not embedded (is pkg/wasmhv/wasmbin/wasm-visor.wasm.gz committed?)")
 	}
 	b, err := Get()
 	if err != nil {
@@ -14,7 +15,7 @@ func TestEmbeddedGet(t *testing.T) {
 	}
 	// the wasm-visor binary starts with the wasm magic \0asm
 	if len(b) < 4 || string(b[:4]) != "\x00asm" {
-		t.Fatalf("not a wasm binary (len=%d, magic=%x)", len(b), b[:min(4, len(b))])
+		t.Fatalf("not a wasm binary (len=%d)", len(b))
 	}
-	t.Logf("decompressed wasm-visor: %.1f MB", float64(len(b))/1024/1024)
+	t.Logf("embedded wasm-visor: %.1f MB decompressed", float64(len(b))/1024/1024)
 }
